feat: add reservations CRUD

+++ apps/ytraveling-backend/src/modules/reservations/reservations.service.ts
@@ -1,20 +1,41 @@
-import { Injectable, InternalServerErrorException, NotFoundException, BadRequestException } from '@nestjs/common';
-import { BookingType } from '@prisma/client';
-import { PrismaService } from '../../tools/database.config';
-import { CreateReservationDto, UpdateReservationDto } from './dto/reservation.dto';
+import {
+  Injectable,
+  InternalServerErrorException,
+  NotFoundException,
+  BadRequestException,
+} from '@nestjs/common';
+import { PrismaService } from '../../../../../lib/prisma-shared/prisma.service';
+import {
+  CreateReservationDto,
+  UpdateReservationDto,
+  BookingType,
+} from './dto/reservation.dto';
 
 @Injectable()
 export class ReservationService {
-  findAll() {
-      throw new Error('Method not implemented.');
-  }
   constructor(private prisma: PrismaService) {}
 
-  async createReservation(createReservationDto: CreateReservationDto) {
-    try {
-      const { userId, roomId, transportId, paymentId, eventId, bookingType, ...rest} = createReservationDto;
+  private validateDate(startDate: Date, endDate: Date) {
+    if (startDate > endDate) {
+      throw new BadRequestException('startDate must be before endDate');
+    }
+  }
 
-      if (![BookingType.ROOM, BookingType.EVENT, BookingType.TRANSPORT].includes(bookingType)) {
+  async createReservation(createReservationDto: CreateReservationDto) {
+    this.validateDate(
+      createReservationDto.startDate,
+      createReservationDto.endDate,
+    );
+
+    try {
+      const { userId, roomId, transportId, eventId, bookingType, ...rest } =
+        createReservationDto;
+
+      if (
+        ![BookingType.ROOM, BookingType.EVENT, BookingType.TRANSPORT].includes(
+          bookingType,
+        )
+      ) {
         throw new BadRequestException('Invalid bookingType');
       }
 
@@ -27,42 +48,49 @@ export class ReservationService {
         }
         case BookingType.EVENT: {
           if (!eventId) {
-            throw new BadRequestException('eventId is required for EVENT bookingType');
+            throw new BadRequestException(
+              'eventId is required for EVENT bookingType',
+            );
           }
           toot = { event: { connect: { id: eventId } } };
           break;
         }
         case BookingType.TRANSPORT: {
           if (!transportId) {
-            throw new BadRequestException('transportId is required for TRANSPORT bookingType');
+            throw new BadRequestException(
+              'transportId is required for TRANSPORT bookingType',
+            );
           }
           toot = { transport: { connect: { id: transportId } } };
           break;
         }
       }
 
-      const reservation = await this.prisma.reservation.create({
+      await this.prisma.reservation.create({
         data: {
           user: { connect: { id: userId } },
           ...toot,
-          ...rest
-        }
+          ...rest,
+          bookingType: bookingType,
+        },
       });
 
-      return reservation;
+      return 'Reservation created successfully';
     } catch (error) {
       console.error(error);
       throw new InternalServerErrorException('Internal Server Error');
     }
   }
 
-  async getReservations() {
+  async getAllReservations() {
     try {
       const reservations = await this.prisma.reservation.findMany();
+
       if (!reservations || reservations.length === 0) {
-        return { status: 204, message: 'No reservations found' };
+        throw new NotFoundException('No reservations found');
       }
-      return { status: 200, message: 'Reservations found', data: reservations };
+
+      return reservations;
     } catch (error) {
       console.error(error);
       throw new InternalServerErrorException('Internal Server Error');
@@ -71,24 +99,74 @@ export class ReservationService {
 
   async getReservationById(id: number) {
     try {
-      const reservation = await this.prisma.reservation.findUnique({ where: { id } });
+      const reservation = await this.prisma.reservation.findUnique({
+        where: { id },
+      });
+
       if (!reservation) {
-        throw new NotFoundException("Reservation not found");
+        throw new NotFoundException('No reservation found');
       }
-      return { status: 200, message: 'Reservation found', data: reservation };
+
+      return reservation;
     } catch (error) {
       console.error(error);
       throw new InternalServerErrorException('Internal Server Error');
     }
   }
 
-  async updateReservation(reservationId: number, updateReservationDto: UpdateReservationDto) {
+  async updateReservation(
+    reservationId: number,
+    updateReservationDto: UpdateReservationDto,
+  ) {
     try {
-      const reservation = await this.prisma.reservation.update({
+      const { userId, roomId, transportId, eventId, bookingType, ...rest } =
+        updateReservationDto;
+
+      if (
+        ![BookingType.ROOM, BookingType.EVENT, BookingType.TRANSPORT].includes(
+          bookingType,
+        )
+      ) {
+        throw new BadRequestException('Invalid bookingType');
+      }
+
+      let toot;
+
+      switch (bookingType) {
+        case BookingType.ROOM: {
+          toot = { room: { connect: { id: roomId } } };
+          break;
+        }
+        case BookingType.EVENT: {
+          if (!eventId) {
+            throw new BadRequestException(
+              'eventId is required for EVENT bookingType',
+            );
+          }
+          toot = { event: { connect: { id: eventId } } };
+          break;
+        }
+        case BookingType.TRANSPORT: {
+          if (!transportId) {
+            throw new BadRequestException(
+              'transportId is required for TRANSPORT bookingType',
+            );
+          }
+          toot = { transport: { connect: { id: transportId } } };
+          break;
+        }
+      }
+
+      await this.prisma.reservation.update({
         where: { id: reservationId },
-        data: updateReservationDto,
+        data: {
+          user: { connect: { id: userId } },
+          ...toot,
+          ...rest,
+        },
       });
-      return reservation;
+
+      return 'Reservation updated successfully';
     } catch (error) {
       console.error(error);
       throw new InternalServerErrorException('Internal Server Error');
@@ -97,11 +175,17 @@ export class ReservationService {
 
   async deleteReservation(id: number) {
     try {
-      const deletedReservation = await this.prisma.reservation.delete({ where: { id } });
-      if (!deletedReservation) {
-        return { status: 204, message: 'No reservation found' };
+      const reservation = await this.prisma.reservation.findUnique({
+        where: { id },
+      });
+
+      if (!reservation) {
+        throw new NotFoundException('No reservation found');
       }
-      return { status: 200, message: 'Reservation deleted', data: deletedReservation };
+
+      await this.prisma.reservation.delete({ where: { id } });
+
+      return 'Reservation deleted successfully';
     } catch (error) {
       console.error(error);
       throw new InternalServerErrorException('Internal Server Error');
